requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic_core==2.27.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==1.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.32.1

